requirements.txt
alembic==1.17.0
altair==5.5.0
annotated-types==0.7.0
APScheduler==3.11.0
attrs==25.4.0
black==25.9.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
narwhals==2.8.0
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
protobuf==6.32.1
psycopg2==2.9.11
pyarrow==21.0.0
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
schedule==1.2.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.44
streamlit==1.50.0
structlog==25.4.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
